New module for input type
